# Test tidy should not count untracked paths towards entries limit