Fixes #103951
r? @jyn514