@Nilstrieb 2022-11-04
What's the perf impact of this? Could you run `time x test tidy` before and after on your machine?