@chenyukang 2022-11-04
@Nilstrieb 

Old:
```
cat@oops:~/rust$ hyperfine --runs 5 'x test tidy'
Benchmark 1: x test tidy
  Time (mean ± σ):      8.154 s ±  1.040 s    [User: 24.769 s, System: 3.725 s]
  Range (min … max):    7.140 s …  9.679 s    5 runs
```

New:
```
cat@oops:~/rust$ hyperfine --runs 5 'x test tidy'
Benchmark 1: x test tidy
  Time (mean ± σ):      8.185 s ±  1.048 s    [User: 26.050 s, System: 4.057 s]
  Range (min … max):    7.370 s …  9.952 s    5 runs
```