@chenyukang 2022-11-04
> Thanks! I'm surprised how simple this is, nice work :) have you tested locally that it ignores untracked files in src/test/ui?

Yes, I tested it locally.
I'm curious how this crate find out the `.gitignore`, it will find all the parents of a file/dir, 
https://github.com/BurntSushi/ripgrep/blob/000015791742bb1280f1853adb714fdee1ba9f8e/crates/ignore/src/dir.rs#L195

```console
add path: "/home/cat/rust/.gitignore"
dir_for_ignorefile: "/home/cat/rust/.git"
add path: "/home/cat/rust/.git/info/exclude"
dir_for_ignorefile: "/home/cat/rust/src"
dir_for_ignorefile: "/home/cat/rust/src"
dir_for_ignorefile: "/home/cat/rust/src/.git"
dir_for_ignorefile: "/home/cat/rust/src/test"
dir_for_ignorefile: "/home/cat/rust/src/test"
dir_for_ignorefile: "/home/cat/rust/src/test/.git"
dir_for_ignorefile: "/home/cat/rust/src/test/ui"
dir_for_ignorefile: "/home/cat/rust/src/test/ui"
```

But the performance seems OK.